#firma sayfası

+++ backend/users/models.py
@@ -153,7 +153,7 @@ class sube(models.Model):
     firma_defter_turu = models.CharField(max_length=100,blank=True, null=True,verbose_name="Firma Defter Türü ",choices=firma_defter_turu_tutma,default="İşletme")
     mukellefiyet_turu = models.CharField(max_length=100,blank=True, null=True,verbose_name="Mükellefiyet Türü ",choices=defter_mukelefiyet_turu_secme,default="Gerçek Kişi")
     stok_bilgisi = models.CharField(max_length=100,blank=True, null=True,verbose_name="Stok Envanter Bilgileri",choices=stok_envanter_bilgileri_secme,default="FİFO")
-    gecici_vergi_orani = models.BigIntegerField(blank=True, null=True,verbose_name="Geçici Vergi Oranı %")
+    gecici_vergi_orani = models.FloatField(blank=True, null=True,verbose_name="Geçici Vergi Oranı %")
     silinme_bilgisi = models.BooleanField(default=False)
